create project folder
