+++ vqgan_jax/configuration_vqgan.py
